+++ GuaranteedRate.Sextant/Metrics/Datadog/DatadogReporter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using GuaranteedRate.Sextant.Config;
 using GuaranteedRate.Sextant.WebClients;
+using Metrics;
 using Newtonsoft.Json;
 
 namespace GuaranteedRate.Sextant.Metrics.Datadog
@@ -22,15 +23,16 @@ namespace GuaranteedRate.Sextant.Metrics.Datadog
         public static string DATADOG_QUEUE_SIZE = "DatadogReporter.QueueSize";
         public static string DATADOG_RETRY_LIMIT = "DatadogReporter.RetryLimit";
         public static string DATADOG_TAGS = "DatadogReporter.Tags";
+        public static string DATADOG_TRACK_HOSTMACHINE = "DatadogReporter.TrackHostmachine";
 
         #endregion
 
         public DatadogReporter(IEncompassConfig config)
-            : base(CreateUrl(config.GetValue(DATADOG_URL),config.GetValue(DATADOG_APIKEY)),
+            : base(CreateUrl(config.GetValue(DATADOG_URL), config.GetValue(DATADOG_APIKEY)),
                 config.GetValue(DATADOG_QUEUE_SIZE, 1000),
                 config.GetValue(DATADOG_RETRY_LIMIT, 3))
         {
-            Setup();
+            Setup(config.GetValue(DATADOG_TRACK_HOSTMACHINE, true));
             var tags = config.GetValue(DATADOG_TAGS, "");
             if (!string.IsNullOrEmpty(tags))
             {
@@ -43,10 +45,10 @@ namespace GuaranteedRate.Sextant.Metrics.Datadog
             }
         }
 
-        public DatadogReporter(string endpoint, string apiKey, int queueSize = 1000, int retries = 3)
-            :base(CreateUrl(endpoint, apiKey), queueSize, retries)
+        public DatadogReporter(string endpoint, string apiKey, int queueSize = 1000, int retries = 3, bool reporthost = true)
+            : base(CreateUrl(endpoint, apiKey), queueSize, retries)
         {
-            Setup();
+            Setup(reporthost);
         }
 
         protected static string CreateUrl(string url, string apiKey)
@@ -54,9 +56,11 @@ namespace GuaranteedRate.Sextant.Metrics.Datadog
             return $"{url}?api_key={apiKey}";
         }
 
-        private void Setup()
+        private void Setup(bool reporthost)
         {
-            _host = Environment.MachineName;
+            _host = reporthost
+                ? Environment.MachineName
+                : null;
             jsonTags = new List<string>();
         }
 
@@ -116,7 +120,39 @@ namespace GuaranteedRate.Sextant.Metrics.Datadog
 
             e.points = points;
 
-            var s = new Series {series = new List<Event> {e}};
+            var s = new Series { series = new List<Event> { e } };
+
+            ReportEvent(s);
+        }
+
+        public void AddHealthCheck(string name, Func<HealthCheckResult> check)
+        {
+            var e = new Event
+            {
+                metric = name,
+                type = "healthcheck",
+                host = _host,
+                tags = jsonTags
+            };
+
+            var result = check.Invoke();
+
+            e.points = new List<IList<long>>() {
+                new []
+                {
+                    GetEpochTime(),
+                    result.IsHealthy == true ? 1 : 0
+                }
+            };
+
+            e.values = new List<IList<string>>() {
+                new []
+                {
+                    result.Message
+                }
+            };
+
+            var s = new Series { series = new List<Event> { e } };
 
             ReportEvent(s);
         }
@@ -126,6 +162,7 @@ namespace GuaranteedRate.Sextant.Metrics.Datadog
             //Using lowercase to help with json
             public string metric { get; set; }
             public IList<IList<long>> points { get; set; }
+            public IList<IList<string>> values { get; set; }
             public string type { get; set; }
             public string host { get; set; }
             public IList<string> tags { get; set; }